backend/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.44.0
sqlalchemy==2.0.49
psycopg2-binary==2.9.11
alembic==1.18.4
pydantic[email]==2.13.1
python-dotenv==1.2.2
groq==1.1.2
python-dateutil==2.9.0.post0
